refactor(core): split wps shape and textbox projection out of drawing-projection
drawing-projection.ts crossed the max-lines lint cap; the pure direct-child
reads (vector shape, textbox story, shared EMU/child helpers) move to
drawing-shape-projection.ts with types re-exported from their original path.
